make interface and move base methods?
