+++ app/Libraries/ChangeUsername.php
@@ -13,7 +13,6 @@ class ChangeUsername
     use Validatable;
 
     const LESS_VALIDATION_TYPES = ['admin', 'revert', 'support'];
-    const MAX_FIELD_LENGTHS = [];
 
     protected $type;
