+++ app/Libraries/Fulfillments/OrderFulfiller.php
@@ -13,8 +13,6 @@ abstract class OrderFulfiller implements Fulfillable
 {
     use Validatable;
 
-    const MAX_FIELD_LENGTHS = [];
-
     const TAGGED_NAME = 'order-fulfiller';
 
     /** @var Order */
