+++ app/Libraries/Payments/PaymentProcessor.php
@@ -21,8 +21,6 @@ abstract class PaymentProcessor implements \ArrayAccess
 {
     use Memoizes, Validatable;
 
-    const MAX_FIELD_LENGTHS = [];
-
     protected $params;
     protected $signature;
