+++ app/Models/Forum/TopicPoll.php
@@ -13,8 +13,6 @@ class TopicPoll
 {
     use Validatable;
 
-    const MAX_FIELD_LENGTHS = [];
-
     private $topic;
     private $validated = false;
     private $params;
