+++ app/Models/Model.php
@@ -10,18 +10,18 @@ use App\Libraries\MorphMap;
 use App\Libraries\Transactions\AfterCommit;
 use App\Libraries\Transactions\AfterRollback;
 use App\Libraries\TransactionStateManager;
+use App\Libraries\ValidatableInterface;
 use App\Scopes\MacroableModelScope;
+use App\Traits\Validatable;
 use Carbon\Carbon;
 use Exception;
 use Illuminate\Database\ClassMorphViolationException;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model as BaseModel;
 
-abstract class Model extends BaseModel
+abstract class Model extends BaseModel implements ValidatableInterface
 {
-    use HasFactory;
-
-    const MAX_FIELD_LENGTHS = [];
+    use HasFactory, Validatable;
 
     protected $connection = 'mysql';
     protected $guarded = [];
@@ -193,9 +193,7 @@ abstract class Model extends BaseModel
             $result = $this->save($options);
 
             if ($result === false) {
-                $message = method_exists($this, 'validationErrors') ?
-                    $this->validationErrors()->toSentence() :
-                    'failed saving model';
+                $message = $this->validationErrors()?->toSentence() ?? 'failed saving model';
 
                 throw new ModelNotSavedException($message);
             }
@@ -214,6 +212,24 @@ abstract class Model extends BaseModel
     public function tableName(bool $includeDbPrefix = false)
     {
         return ($includeDbPrefix ? $this->dbName().'.' : '').$this->getTable();
+    }
+
+    public function validateDbFieldLength(int $limit, string $dbField, ?string $checkField = null): void
+    {
+        if ($this->isDirty($dbField)) {
+            $this->validateFieldLength($limit, $dbField, $checkField);
+        }
+    }
+
+    public function validateDbFieldLengths(): void
+    {
+        foreach (static::MAX_FIELD_LENGTHS as $field => $limit) {
+            $this->validateDbFieldLength($limit, $field, $field);
+        }
+    }
+
+    public function validationErrorsTranslationPrefix() {
+        return '';
     }
 
     /**
